new_ones/backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0

# Database
pymongo==4.6.1

# Security
Werkzeug==3.0.1
cryptography==41.0.7

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2

# Network Monitoring
pyshark==0.6
scapy==2.5.0

# Data Processing
python-dateutil

# HTTP Client
requests

# Utilities
click
colorama

# Development
pytest
pytest-cov
black
flake8